+++ changelog.md
@@ -1,3 +1,5 @@
-v0.13.0-a2:
-- added `call` instr to MIR
-- typedefs and datadefs are now being translated
+v0.13.0-a3:
+- Added `And`, `Or`, `Xor` bin ops to MIR
+- Added `Neg` arithmetical instruction to MIR
+- Added `Load`, `Store`, `GetAddr` memory instrs 
+- fixes